完成了box的显示，现在显示异常，需要查找原因

+++ DXLearn/DXLearn/Common/GameTimer.cpp
@@ -25,7 +25,7 @@ float GameTimer::TotalTime() const
 
 float GameTimer::DeltaTime() const
 {
-    return mDeltaTime;
+    return static_cast<float>(mDeltaTime);
 }
 
 void GameTimer::Reset()
